+++ kernel/src/main/java/kernel/structural/laws/Law.java
@@ -4,5 +4,5 @@ import kernel.Measurement;
 import kernel.NamedElement;
 
 public interface Law extends NamedElement {
-    Measurement generateNextMeasurement(int t);
+    Measurement generateNextMeasurement(double t);
 }
